残り問題数が0の時に起こるバグを修正

+++ models/grade.go
@@ -1,6 +1,6 @@
 package models
 
 type Grade struct {
-	Solvable int `json:"solvable"`
-	All      int `json:"all"`
+	Solvable int64 `json:"solvable"`
+	All      int   `json:"all"`
 }
